requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pandas>=1.2.0
platformdirs>=2.0.0
tkinterdnd2>=0.3.0; platform_system != "Linux"
